Move existing metadata types to new Metadata sub-folder/sub-namespace
