+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Node/JsonNodeConverterFactory.cs
@@ -3,6 +3,7 @@
 
 using System.Diagnostics;
 using System.Text.Json.Node;
+using System.Text.Json.Serialization.Metadata;
 
 namespace System.Text.Json.Serialization.Converters
 {
