+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Metadata/MemberAccessor.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using System.Reflection;
 
-namespace System.Text.Json
+namespace System.Text.Json.Serialization.Metadata
 {
     internal abstract class MemberAccessor
     {
