+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Metadata/ParameterRef.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace System.Text.Json
+namespace System.Text.Json.Serialization.Metadata
 {
     internal readonly struct ParameterRef
     {
