+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Metadata/ReflectionEmitMemberAccessor.cs
@@ -8,7 +8,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using System.Reflection.Emit;
 
-namespace System.Text.Json.Serialization
+namespace System.Text.Json.Serialization.Metadata
 {
     internal sealed class ReflectionEmitMemberAccessor : MemberAccessor
     {
