+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/WriteStackFrame.cs
@@ -5,6 +5,7 @@ using System.Collections;
 using System.Diagnostics;
 using System.Threading.Tasks;
 using System.Text.Json.Serialization;
+using System.Text.Json.Serialization.Metadata;
 
 namespace System.Text.Json
 {
